pyproject.toml
[project]
name = "agenticai"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "fastapi[all]>=0.124.0",
    "openai-agents>=0.6.2",
    "python-dotenv>=1.2.1",
]
